feat: add category api get

+++ back-end/trendify/src/main/java/com/weiz/trendify/service/dto/request/product/ProductCreateDto.java
@@ -23,5 +23,5 @@ public class ProductCreateDto {
     MultipartFile imageFile;
 
     @NotNull
-    Long categoryId;
+    Integer categoryId;
 }
